# contactify-quark/src/main/java/org/acme/contactify/account/AuthResponseDTO.java
package org.acme.contactify.account;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter @Setter @AllArgsConstructor @NoArgsConstructor
public class AuthResponseDTO {
    private String username;
    private String authToken;
}