# contactify-react/docs/openapiSpecs.yaml
...
openapi: 3.0.3
info:
  title: contactify-quark API
  version: 1.0.0-SNAPSHOT
paths:
...
  /user/login:
    post:
      tags:
      - Account Resource
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequestDTO'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponseDTO'
...
  /user/register:
    post:
      tags:
      - Account Resource
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
components:
  schemas:
    Account:
      type: object
      properties:
        id:
          format: int64
          type: integer
        username:
          type: string
        password:
          type: string
        roles:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/AccountRole'
    AccountRole:
      type: object
      properties:
        id:
          format: int64
          type: integer
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        role:
          type: string
    AuthRequestDTO:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponseDTO:
      type: object
      properties:
        username:
          type: string
        authToken:
          type: string
...
  securitySchemes:
    SecurityScheme:
      type: http
      description: Authentication
      scheme: basic